Create a while loop that takes input from the user
This is a notepad program created for Ex. 2 of Loops & Iterators:
https://launchschool.com/books/ruby/read/loops_iterators

"Write a while loop that takes input from the user, performs an action,
and only stops when the user types "STOP".
Each loop can get info from the user."

This notepad program takes input(notes) from the user and writes them
to a file and continues to do so while the input is not 'STOP'.
If the user types 'STOP', the program will close the notepad and then
display the notes to the user.