Step 7 - Change Font Size and Color
Change the font size on the entire page by using a CSS element
selector for the body element. Because the font-size property is
inherited, all of the text on the page will be resized to be relative
to this new font-size. Change the size to 20px. After you've viewed
the change in font size, add another CSS property to the body element
selector that changes all of the text color to #666666.

https://launchschool.com/lessons/05a74171/assignments/4233