Step 9 - Add Class Selectors
reate two new CSS class selectors, one to use for the ingredients
paragraphs and the other to use for the directions paragraphs.
Do not use the same names as the id attributes already used.
For ingredient paragraphs, set the font-style property to italic.
For direction paragraphs, set their font-family to sans-serif.

https://launchschool.com/lessons/05a74171/assignments/4233